requirements_ia.txt
# IA y Gemini
google-generativeai
python-dotenv

# Core
streamlit
pandas
numpy
requests
